+++ djangoVersion/posts/static/sell-platform.css
@@ -288,13 +288,7 @@
     margin-bottom: 10px;
     text-align: left;
   }
-  
-  .newItem p {
-    font-size: 20px;
-    margin-bottom: 10px;
-    text-align: left;
-    color: #ff6347;
-  }
+
   
   .newItem button {
     background-color: #ff7f50;
@@ -325,6 +319,14 @@
 
   .item-text-container {
     width: 100%;
+  }
+
+  .item-text-container p{
+    width: 100%;
+    font-size: 20px;
+    margin-bottom: 10px;
+    text-align: left;
+    color: #ff6347;
   }
 
 /* 商品詳細資料 */
@@ -396,5 +398,17 @@
   background-color: rgb(252, 89, 56);
   color: white;
 }
+
+.comment-content{
+  font-size: 16px;
+  color: black;
+}
+
+.comment-time{
+  font-size: 12px;
+  color: #888;
+}
+
+
