+++ djangoVersion/posts/templates/item_detail.html
@@ -1,56 +1,118 @@
-{% extends 'index.html' %}
+{% load static %}
 
-{% block content %}
-    <div class="newItem">
-        <div class="item-image-container">
-            <img src="{{ item.image.url }}" alt="{{ item.name }}"></img>
+<!DOCTYPE html>
+<html lang="zh-TW">
+<head>
+  <meta charset="UTF-8">
+  <title>My Web App</title>
+  <link rel="stylesheet" href="{% static 'sell-platform.css' %}?v=1">
+</head>
+<body style="background-color: #faeccd;">
+    
+    <header class="header-container1">
+        <div class="header-links1">
+            <a href="#" class="header-link1" style="margin-left: 40px;">語言</a>
+            <a href="#" class="header-link1">登入</a>
+            <a href="#" class="header-link1">註冊</a>
         </div>
-        <div class="item-text-container">
-            <h3>{{ item.name }}</h3>
-            <p>＄{{ item.price }}</p>
-            <a href="{% url 'item_details' item_id=item.pk %}" class="item-details-button">詳細資訊</a>
+    </header>
+    <header class="header-container2">
+
+        <h1 class="header-logo">出清台大2.0</h1>
+        <div class="header-links2">
+            <a href="../homePage/mainPage.html" class="header-link2" style="margin-left: 150px;">主頁</a>
+            <a href="#" class="header-link2">活動區</a>
+            <a href="#" class="header-link2">二手物品交易區</a>
+            <a href="#" class="header-link2" style="margin-right: 190px;">聊天室</a>
+        </div>
+        <div class="header-links1">
+            <div class="search-bar">
+                <input type="text" placeholder="搜尋...">
+            </div>
         </div>
     </div>
+    </header>
 
-    <div class="item-details-wrapper">
-        <div class="item-details-container">
-            <!-- 商品詳細資訊 -->
-            <div class="item-details-image">
-                <img src="{{ item.image.url }}" alt="{{ item.name }}" />
-            </div>
-            <div class="item-details-text">
-                <h2>{{ item.name }}</h2>
-                <h3>＄{{ item.price }}</h3>
-                <p>數量：{{ item.quantity }}</p>
-                <p>分類：{{ item.category }}</p>
-                <p>交易方式：{{ item.transaction_type }}</p>
-                <p>面交地點：{{ item.transaction_location }}</p>
-                <p>商品狀況：{{ item.condition }}</p>
-            </div>
+    <hr style="border: 0.1px solid black; margin: 0;">
 
-            <label>留言板：</label>
-            <form method="post" action="{% url 'item_detail' item_id=item.pk %}" id="comment-form">
-                {% csrf_token %}
-                <textarea id="comment-input" name="comment" rows="5" cols="50"></textarea>
-                <br>
-                <button type="submit" id="comment-submit-button">留言</button>
-            </form>
+    <!-- 按鈕 -->
+    <button id="sell-button">我要販賣物品</button>
+    
+    <!-- 商品類別 -->
+    <div class="categories">
+        <ul>
+            <li class="{% if not category %}active{% endif %}">
+                <a href="{% url 'item_list' %}" onclick="location.href=this.href;">全部</a>
+            </li>
+            <li class="{% if category == '衣物' %}active{% endif %}">
+                <a href="{% url 'item_list_by_category' '衣物' %}" onclick="location.href=this.href;">衣物</a>
+            </li>
+            <li class="{% if category == '鞋子' %}active{% endif %}">
+                <a href="{% url 'item_list_by_category' '鞋子' %}" onclick="location.href=this.href;">鞋子</a>
+            </li>
+            <li class="{% if category == '3C' %}active{% endif %}">
+                <a href="{% url 'item_list_by_category' '3C' %}" onclick="location.href=this.href;">3C</a>
+            </li>
+            <li class="{% if category == '其他' %}active{% endif %}">
+                <a href="{% url 'item_list_by_category' '其他' %}" onclick="location.href=this.href;">其他</a>
+            </li>
+        </ul>
+    </div>
 
-            <div class="comments">
-                {% for comment in comments %}
-                    <div class="comment">
-                        <p>{{ comment.content }}</p>
-                        <p>{{ comment.created_at }}</p>
-                    </div>
-                {% empty %}
-                    <p>目前還沒有留言。</p>
-                {% endfor %}
-            </div>
+    <div id="item-container">
+        {% for item in items %}
+            <div class="newItem">
+                <div class="item-image-container">
+                    <img src="{{ item.image.url }}" alt="{{ item.name }}"></img>
+                </div>
+                <div class="item-text-container">
+                    <h3>{{ item.name }}</h3>
+                    <p>＄{{ item.price }}</p>
+                    <button class="item-details-button" data-item-id="{{ item.item_id }}">詳細資訊</button>
+                </div>
+                {% comment %} <div class="item-details-wrapper"> {% endcomment %}
+                    <div class="item-details-container" data-item-id="{{ item.item_id }}">
+                        <!-- 商品詳細資訊 -->
+                        <div class="item-details-image">
+                            <img src="{{ item.image.url }}" alt="{{ item.name }}" />
+                        </div>
+                        <div class="item-details-text">
+                            <h2>{{ item.name }}</h3>
+                            <h3>＄{{ item.price }}</p>
+                            <p>數量：{{ item.quantity }}</p>
+                            <p>分類：{{ item.category }}</p>
+                            <p>交易方式：{{ item.transaction_type }}</p>
+                            <p>面交地點：{{ item.transaction_location }}</p>
+                            <p>商品狀況：{{ item.condition }}</p>
+                        </div>
 
-            <button class="close-button">X</button>
+                        <label>留言板：</label>
+                        <form method="post" action="{% url 'item_detail' item_id=item.pk %}" id="comment-form">
+                            {% csrf_token %}
+                            <textarea id="comment-input" name="comment" rows="5" cols="50"></textarea>
+                            <br>
+                            <button type="submit" id="comment-submit-button">留言</button>
+                        </form>
+                    
+                        <div class="comments">
+                            {% for comment in comments %}
+                                <div class="comment">
+                                    <p class="comment-content">{{ comment.content }}</p>
+                                    <p class="comment-time">{{ comment.created_at }}</p>
+                                </div>
+                            {% empty %}
+                                <p>目前還沒有留言。</p>
+                            {% endfor %}
+                        </div>
+                    
+                        <button class="close-button" data-item-id="{{ item.item_id }}">X</button>
                         
+                    </div>
+                {% comment %} </div> {% endcomment %}
             </div>
-        {% comment %} </div> {% endcomment %}
+        {% endfor %}
     </div>
-</div>
-{% endblock %}
+    
+    <script src="{% static 'item-detail.js' %}"></script>
+</body>
+</html>
